pyproject.toml
[project]
name = "ray-serve-demo"
version = "0.1.0"
readme = "README.md"
requires-python = ">=3.7"
dependencies = [
    "mlflow",
    "ipykernel",
    "virtualenv",
    "python-dotenv",
    "ray[serve]",
    "pip",
    "pydantic",
    "starlette",
    "boto3",
]

[tool.setuptools.packages.find]
include = ["mlray"]

[tool.uv]
package = true

[project.scripts]
mlray = "mlray.cli.main:main"
